+++ cli/cmdcontext/cmdcontext.go
@@ -85,4 +85,6 @@ type CliCtx struct {
 	TarantoolCli TarantoolCli
 	// IntegrityCheck is a public key used for integrity check.
 	IntegrityCheck string
+  // This flag disables searching of other tt versions to run instead of the current one.
+	IsSelfExec bool
 }